Container runtime structure refactoring
